+++ src/components/popconfirm/index.ts
@@ -1,0 +1,2 @@
+export { MPPopconfirm } from './MPPopconfirm';
+export type { MPPopconfirmProps } from './MPPopconfirm';