+++ src/index.ts
@@ -73,6 +73,7 @@ export * from './components/page-layout';
 export * from './components/pagination';
 export * from './components/panes';
 export * from './components/pill';
+export * from './components/popconfirm';
 export * from './components/popover';
 export * from './components/progress-box';
 export * from './components/progress-circular';